Plugins updates

+++ plugins/google-site-kit/includes/Core/Authentication/Authentication.php
@@ -277,6 +277,7 @@ final class Authentication implements Provides_Feature_Metrics {
 		$this->verification_file()->register();
 		$this->verification_meta()->register();
 		$this->has_connected_admins->register();
+		$this->has_multiple_admins->register();
 		$this->owner_id->register();
 		$this->connected_proxy_url->register();
 		$this->disconnected_reason->register();
@@ -311,9 +312,7 @@ final class Authentication implements Provides_Feature_Metrics {
 				}
 
 				$this->set_connected_proxy_url();
-			},
-			10,
-			3
+			}
 		);
 
 		add_filter(
@@ -330,7 +329,6 @@ final class Authentication implements Provides_Feature_Metrics {
 				}
 
 				$user['connectURL']           = esc_url_raw( $this->get_connect_url() );
-				$user['hasMultipleAdmins']    = $this->has_multiple_admins->get();
 				$user['initialVersion']       = $this->initial_version->get();
 				$user['isUserInputCompleted'] = ! $this->user_input->are_settings_empty();
 				$user['verified']             = $this->verification->has();
@@ -1144,24 +1142,57 @@ final class Authentication implements Provides_Feature_Metrics {
 						?>
 						<a
 							href="#"
-							onclick="clearSiteKitAppStorage()"
+							onclick="reauthenticateAndContinueSetup()"
 						><?php esc_html_e( 'Click here', 'google-site-kit' ); ?></a>
 					</p>
 					<?php
 					BC_Functions::wp_print_inline_script_tag(
 						sprintf(
 							"
-							function clearSiteKitAppStorage() {
-								if ( localStorage ) {
-									localStorage.clear();
+							function reauthenticateAndContinueSetup() {
+								const moduleSlug = getAbandonedModuleSlug();
+
+								if ( moduleSlug ) {
+									const redirect = '%3\$s&slug=' + moduleSlug;
+									document.location = '%2\$s&redirect=' + encodeURIComponent( redirect );
+								} else {
+									if ( localStorage ) {
+										localStorage.clear();
+									}
+									if ( sessionStorage ) {
+										sessionStorage.clear();
+									}
+									document.location = '%2\$s';
 								}
-								if ( sessionStorage ) {
-									sessionStorage.clear();
+							}
+
+							function getAbandonedModuleSlug() {
+								for ( const storage of [ localStorage, sessionStorage ] ) {
+									if ( ! storage ) {
+										continue;
+									}
+									const key = Object.keys( storage ).find( ( k ) =>
+										k.match( 'googlesitekit_%1\$s_.*_module_setup' )
+									);
+									if ( ! key ) {
+										continue;
+									}
+									try {
+										return JSON.parse( storage[ key ] )?.value;
+									} catch ( _ ) {}
 								}
-								document.location = '%s';
 							}
 							",
-							esc_url_raw( $this->get_connect_url() )
+							GOOGLESITEKIT_VERSION,
+							esc_url_raw( $this->get_connect_url() ),
+							esc_url_raw(
+								$this->context->admin_url(
+									'dashboard',
+									array(
+										'reAuth' => 'true',
+									)
+								)
+							)
 						)
 					);
 					return ob_get_clean();
